// Capstone Project/src/app/pages/admin/admin-post-management-page/screen/types.ts
export enum adminPostManagmentPageEnum {
  GET_POST_ADMIN_REQUEST = 'APP/ADMIN_POST_PAGE/GET_POST_ADMIN_REQUEST',
  GET_POST_ADMIN_SUCCESS = 'APP/ADMIN_POST_PAGE/GET_POST_ADMIN_SUCCESS',
  GET_POST_ADMIN_FAILURE = 'APP/ADMIN_POST_PAGE/GET_POST_ADMIN_FAILURE',

  SEARCH_POST_ADMIN_REQUEST = 'APP/ADMIN_POST_PAGE/SEARCH_POST_ADMIN_REQUEST',
  SEARCH_POST_ADMIN_SUCCESS = 'APP/ADMIN_POST_PAGE/SEARCH_POST_ADMIN_SUCCESS',
  SEARCH_POST_ADMIN_FAILURE = 'APP/ADMIN_POST_PAGE/SEARCH_POST_ADMIN_FAILURE',


  CONFIRM_POST_ADMIN_REQUEST = 'APP/ADMIN_POST_PAGE/CONFIRM_POST_ADMIN_REQUEST',
  CONFIRM_POST_ADMIN_SUCCESS = 'APP/ADMIN_POST_PAGE/CONFIRM_POST_ADMIN_SUCCESS',
  CONFIRM_POST_ADMIN_FAILURE = 'APP/ADMIN_POST_PAGE/CONFIRM_POST_ADMIN_FAILURE',

  REJECT_POST_ADMIN_REQUEST = 'APP/ADMIN_POST_PAGE/REJECT_POST_ADMIN_REQUEST',
  REJECT_POST_ADMIN_SUCCESS = 'APP/ADMIN_POST_PAGE/REJECT_POST_ADMIN_SUCCESS',
  REJECT_POST_ADMIN_FAILURE = 'APP/ADMIN_POST_PAGE/REJECT_POST_ADMIN_FAILURE',

  DELETE_POST_ADMIN_REQUEST = 'APP/ADMIN_POST_PAGE/DELETE_POST_ADMIN_REQUEST',
  DELETE_POST_ADMIN_SUCCESS = 'APP/ADMIN_POST_PAGE/DELETE_POST_ADMIN_SUCCESS',
  DELETE_POST_ADMIN_FAILURE = 'APP/ADMIN_POST_PAGE/DELETE_POST_ADMIN_FAILURE',

  VERIFY_POST_ADMIN_REQUEST = 'APP/ADMIN_POST_PAGE/VERIFY_POST_ADMIN_REQUEST',
  VERIFY_POST_ADMIN_SUCCESS = 'APP/ADMIN_POST_PAGE/VERIFY_POST_ADMIN_SUCCESS',
  VERIFY_POST_ADMIN_FAILURE = 'APP/ADMIN_POST_PAGE/VERIFY_POST_ADMIN_FAILURE',

  CANCEL_VERIFY_POST_ADMIN_REQUEST = 'APP/ADMIN_POST_PAGE/CANCEL_VERIFY_POST_ADMIN_REQUEST',
  CANCEL_VERIFY_POST_ADMIN_SUCCESS = 'APP/ADMIN_POST_PAGE/CANCEL_VERIFY_POST_ADMIN_SUCCESS',
  CANCEL_VERIFY_POST_ADMIN_FAILURE = 'APP/ADMIN_POST_PAGE/CANCEL_VERIFY_POST_ADMIN_FAILURE',

  RESTORE_POST_ADMIN_REQUEST = 'APP/ADMIN_POST_PAGE/RESTORE_POST_ADMIN_REQUEST',
  RESTORE_POST_ADMIN_SUCCESS = 'APP/ADMIN_POST_PAGE/RESTORE_POST_ADMIN_SUCCESS',
  RESTORE_POST_ADMIN_FAILURE = 'APP/ADMIN_POST_PAGE/RESTORE_POST_ADMIN_FAILURE',

  CLEAR_STATE = 'APP/ADMIN_POST_PAGE/CLEAR_STATE',
}